cloud-sdk: Use `usize` for number of `probe` retries
We can avoid the use of casting (for u64), which can in fact fail, using `usize`
for the number of probe retries. The `usize` use here seems a good fit
as it matches the word size of the current in-use CPU architecture.

+++ updatehub-cloud-sdk/src/client.rs
@@ -91,7 +91,7 @@ impl<'a> Client<'a> {
 
     pub async fn probe(
         &self,
-        num_retries: u64,
+        num_retries: usize,
         firmware: api::FirmwareMetadata<'_>,
     ) -> Result<api::ProbeResponse> {
         validate_url(self.server)?;
